+++ packages/kit-headless/src/components/select/hidden-select.tsx
@@ -41,10 +41,14 @@ export const HHiddenNativeSelect = component$(
                 // @ts-expect-error modular forms ref function
                 ref?.(element);
               }}
+              onFocus$={() => {
+                // override modular forms focus event
+                return;
+              }}
               multiple={context.multiple}
               tabIndex={-1}
               autocomplete={autoComplete}
-              disabled={context.disabled}
+              disabled={context.isDisabledSig.value ? true : undefined}
               required={context.required}
               name={context.name}
               // height is determined by its children