+++ packages/kit-headless/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.4.4
+
+### Patch Changes
+
+- ✨ new Select.ErrorMessage component (by [@thejackshelton](https://github.com/thejackshelton) in [#825](https://github.com/qwikifiers/qwik-ui/pull/825))
+
+  feat: data-invalid attribute to style when the select is invalid
+
+  feat: new Select.Description component
+
+## 0.4.3
+
+### Patch Changes
+
+- 🐞🩹 select validates correctly with modular forms (by [@thejackshelton](https://github.com/thejackshelton) in [#814](https://github.com/qwikifiers/qwik-ui/pull/814))
+
+- refactor: improved select focus navigation (by [@thejackshelton](https://github.com/thejackshelton) in [#822](https://github.com/qwikifiers/qwik-ui/pull/822))
+
+- 🐞🩹 select can now be reactively disabled (by [@thejackshelton](https://github.com/thejackshelton) in [#823](https://github.com/qwikifiers/qwik-ui/pull/823))
+
 ## 0.4.2
 
 ### Patch Changes